refactor(29): вынести скрипт для header в отдельный файл
