+++ index.html
@@ -195,6 +195,6 @@
 
     </footer>
   </div>
-  <script src="./scripts/index.js"></script>
+  <script src="./scripts/index.js" type="module"></script>
 </body>
 </html>
